+++ falcon/testing/client.py
@@ -72,9 +72,8 @@ class Result(object):
             under Python 2.6 and 2.7, and of type ``str`` otherwise.
             If the content type does not specify an encoding, UTF-8 is
             assumed.
-        json (JSON serializable): Deserialized JSON body. Will be ``None`` if
-            the body has no content to deserialize. Otherwise, raises an error
-            if the response is not valid JSON.
+        json (dict): Deserialized JSON body. Raises ``JSONDecodeError`` if
+            the body is empty or not JSON.
     """
 
     def __init__(self, iterable, status, headers):
@@ -166,9 +165,6 @@ class Result(object):
 
     @property
     def json(self):
-        if not self.text:
-            return None
-
         return util_json.loads(self.text)